+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/generators/proposalActionChangeSettings.ts
@@ -29,8 +29,8 @@ export const generateProposalActionChangeSettings = (
         function: 'settings',
         contract: 'TokenVote',
         parameters: [
-            { type: 'address', value: '0x3f5CE5FBFe3E9af3971dD833D26BA9b5C936F0bE' },
-            { type: 'uint256', value: '1000000000000000000' },
+            { name: 'duration', type: 'uint16', value: 5, notice: 'Changing minimum duration' },
+            { name: 'tokenAmount', type: 'string', value: '1000000000000000000' },
         ],
     },
     ...action,