+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/actions/generators/proposalActionWithdrawToken.ts
@@ -27,8 +27,18 @@ export const generateProposalActionWithdrawToken = (
         function: 'transfer',
         contract: 'Ether',
         parameters: [
-            { type: 'address', value: '0x3f5CE5FBFe3E9af3971dD833D26BA9b5C936F0bE' },
-            { type: 'uint256', value: '1000000000000000000' },
+            {
+                name: 'address',
+                type: 'string',
+                value: '0x3f5CE5FBFe3E9af3971dD833D26BA9b5C936F0bE',
+                notice: 'Contract address of asset',
+            },
+            {
+                name: 'tokenAmount',
+                type: 'string',
+                value: '1000000000000000000',
+                notice: 'Amount of tokens to withdraw',
+            },
         ],
     },
     amount: '10000000',