+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActions/proposalActions.tsx
@@ -1,5 +1,5 @@
 import classNames from 'classnames';
-import { useState, type ReactNode } from 'react';
+import { useRef, useState, type ReactNode } from 'react';
 import { Accordion, Button, Card, Heading } from '../../../../../core';
 import type { IWeb3ComponentProps } from '../../../../types';
 import { useOdsModulesContext } from '../../../odsModulesProvider';
@@ -16,7 +16,7 @@ export interface IProposalActionsProps extends IWeb3ComponentProps {
      */
     actionNames?: Record<string, string>;
     /**
-     * Map of action-type <=> custom-component to customise how actions are displayed.
+     * Map of action-type <=> custom-component to customize how actions are displayed.
      */
     customActionComponents?: Record<string, ProposalActionComponent>;
     /**
@@ -32,9 +32,11 @@ export interface IProposalActionsProps extends IWeb3ComponentProps {
 export const ProposalActions: React.FC<IProposalActionsProps> = (props) => {
     const { actions, actionNames, className, customActionComponents, children, ...web3Props } = props;
 
+    const [expandedItems, setExpandedItems] = useState<string[]>([]);
+
     const { copy } = useOdsModulesContext();
 
-    const [expandedItems, setExpandedItems] = useState<string[]>([]);
+    const actionsContainerRef = useRef<HTMLDivElement | null>(null);
 
     const handleToggleAll = () => {
         if (expandedItems.length === actions.length) {
@@ -42,13 +44,17 @@ export const ProposalActions: React.FC<IProposalActionsProps> = (props) => {
         } else {
             setExpandedItems(Array.from({ length: actions.length }, (_, index) => `${index}`));
         }
+
+        if (actionsContainerRef.current && expandedItems.length === actions.length) {
+            actionsContainerRef.current.scrollIntoView({ behavior: 'instant' });
+        }
     };
 
     const handleAccordionValueChange = (value: string[] = []) => setExpandedItems(value);
 
     return (
         <Card className={classNames('w-full overflow-hidden', className)}>
-            <Heading size="h2" className="px-4 pt-4 md:px-6 md:pt-6">
+            <Heading size="h2" className="p-4 md:p-6" ref={actionsContainerRef}>
                 {copy.proposalActionsContainer.containerName}
             </Heading>
             <Accordion.Container isMulti={true} value={expandedItems} onValueChange={handleAccordionValueChange}>