+++ packages/gov-ui-kit/src/modules/components/proposal/proposalActions/proposalActionsTypes/proposalActionViewMode.ts
@@ -1,0 +1,5 @@
+export enum ProposalActionViewMode {
+    BASIC = 'BASIC',
+    DECODED = 'DECODED',
+    RAW = 'RAW',
+}