+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Add 'Raw' and 'Decoded' views to `ProposalActions` with Dropdown selector for the data of proposal actions with more
+    technical understanding.
+
 ### Fixed
 
 -   Hide minimum participation details on `ProposalVotingBreakdownToken` module component when minParticipation is set